pyproject.toml
[project]
name = "selfhosted-rag-doc-chat-prototype"
version = "0.1.0"
requires-python = ">=3.11, <3.13"  # Fixed: Support 3.12 for Ubuntu 24.04
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-qdrant>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "langchain-chroma>=0.1.4",
    "llama-cpp-python>=0.3.0",
    "fastapi>=0.116.0",
    "uvicorn>=0.35.0",
    "pandas>=3.0.2",
    "duckdb>=1.5.1",
    "transformers>=5.5.0",
    "docling[easyocr]>=2.85.0",
    "langgraph>=0.2.76",
    "mmh3>=4.1.0",
    "pdfplumber>=0.11.0",
    "pdf2image>=1.17.0",
    "pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "pyyaml>=6.0.1",
    "more-itertools>=10.0.0",
    "ftfy>=6.1.1",
    "nltk>=3.8.1",
    "sentence-transformers>=3.0.0",
    "regex>=2023.10.3",
    "beautifulsoup4>=4.12.2",
    "html5lib>=1.1",
    "charset-normalizer>=3.3.0",
    "psutil>=5.9.0",
    "torch>=2.5.1,<2.6.0",
    "torchaudio>=2.5.1,<2.6.0",
    "torchvision>=0.20.1,<0.21.0",
    # Helper libs
    "python-dotenv>=1.1.1",
    "redis>=7.0.0",
    "openai>=1.0.0",
    "protobuf<5.0.0",

    "pydantic>=2.11.0",
    "qdrant-client>=1.16.2",
]